Fix "xref-pop-to-location: Marker does not point anywhere"
Avoid a possible error when jumping to a position in a file that is
out of bounds, due to the jumped-to file containing line number
directives.  E.g., if target.ml's first line is "# 100 \"target.ml\"",
the right place to jump for line 100 would be the second physical line
of the file, but getting this right would be quite complex.  As a
first step, just produce *some* valid position in the right buffer so
that xref-find-definitions does not fail completely.

The ultimate error manifests as:

    xref-pop-to-location: Marker does not point anywhere

(and the xref-find-definitions command jumps to some random buffer):
due to the xref-location-marker method returning a marker that does
not point anywhere.